% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{iddo_palettes}
\alias{iddo_palettes}
\alias{iddo_palettes_discrete}
\alias{iddo_palettes_sequential}
\title{iddo palettes}
\format{
\subsection{\code{iddo_palettes}}{

An object of class \code{palettes_palette} with 14
colour palettes. Use \code{names(iddo_palettes)} to return all palette names.
}

\subsection{\code{iddo_palettes_discrete}}{

An object of class \code{palettes_palette} with
5 discrete colour palettes.
Use \code{names(iddo_palettes_discrete)} to return all palette names.
}

\subsection{\code{iddo_palettes_sequential}}{

An object of class \code{palettes_palette} with
9 sequential colour palettes.
Use \code{names(iddo_palettes_sequential)} to return all palette names.
}
}
\source{
\url{https://github.com/lu-harr/iddoPal}
}
\usage{
iddo_palettes

iddo_palettes_discrete

iddo_palettes_sequential
}
\description{
Palettes for IDDO
}
\examples{
# Get all palettes by name.
names(iddo_palettes)

# Plot all palettes.
plot(iddo_palettes)
}
\seealso{
\code{\link[=pal_palette]{pal_palette()}}, \code{\link[=pal_colour]{pal_colour()}}
}
\author{
\href{https://github.com/lu-harr}{Lucinda Harrison}
}
\keyword{datasets}
