% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iddoPal-package.R
\docType{package}
\name{iddoPal-package}
\alias{iddoPal}
\alias{iddoPal-package}
\title{iddoPal: IDDO Colour Palettes}
\description{
A home for the IDDO colour palette. Built using the 'palettes' package, which provides methods for printing, formatting, casting and coercion, extraction and updating of components, plotting, colour mixing arithmetic, and colour interpolation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luharr/iddoPal}
  \item Report bugs at \url{https://github.com/luharr/iddoPal/issues}
}

}
\author{
\strong{Maintainer}: Lucinda Harrison \email{lucy.harrison@wwarn.org}

}
\keyword{internal}
